refactor: Create TestForm component
